updated bash configs

+++ .bash_aliases
@@ -4,28 +4,38 @@ alias la='ls -al'
 alias l='ls -1'
 
 # batcat
-# alias cat='bat'
+if command -v "batcat" &>/dev/null; then
+  alias cat='bat'
+fi
 
 # Directory
 alias mcd='mkdir'
 
 # Neovim
-alias nv='nvim'
+if command -v "nvim" &>/dev/null; then
+  alias nv='nvim'
+fi
 
 # Lazygit
-alias lg='lazygit'
+if command -v "lazygit" &>/dev/null; then
+  alias lg='lazygit'
+fi
 
 # tmux
-alias ntm='tmux new-session -s '
-alias ktm='tmux kill-server'
-alias tm='tmux'
-alias tma='tmux attach -t '
+if command -v "tmux" &>/dev/null; then
+  alias ntm='tmux new-session -s '
+  alias ktm='tmux kill-server'
+  alias tm='tmux'
+  alias tma='tmux attach -t '
+fi
 
 # git
-alias g='git '
-alias gbr='git br'
-alias gci='git ci'
-alias gco='git co'
-alias ga='git add'
-alias gsw='git sw'
-alias gst='git st'
+if command -v "git" &>/dev/null; then
+  alias g='git '
+  alias gbr='git br'
+  alias gci='git ci'
+  alias gco='git co'
+  alias ga='git add'
+  alias gsw='git sw'
+  alias gst='git st'
+fi
